Added Mikeeeyy  additions
Buildables supported:
WoodRamp_DZ - Allows snapping to the left and right sides.
VaultStorage/VaultStorageLocked - Allows snapping to the left and right
sides as well as the top and bottom.
TentStorage - Allows snapping to the left and right sides.
TentStorageDomed/TentStorageDomed2 - Allows snapping to the left and
right sides.
StorageShed_DZ - Allows snapping to the left and right sides.
WoodFloorHalf_DZ - Allows snapping to the front, back, left and right
sides.
WoodFloorQuarter_DZ - Allows snapping to the front, back, left and right
sides.
WoodStairsSans_DZ/WoodStairs_DZ/WoodStairsRails_DZ - Allows snapping to
the left and right sides.
GunRack_DZ - Allows snapping to the left and right sides as well as the
top and bottom.
WoodCrate_DZ - Allows snapping to the left and right sides as well as
the top and bottom.